requirements.txt
aws-cdk-lib==2.132.0
constructs>=10.0.0,<11.0.0
yfinance==0.2.37
boto3>=1.34.63
